+++ src/main/java/dev/enjarai/restartdetector/block/ModBlocks.java
@@ -7,6 +7,7 @@ import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityTypeBuilder;
+import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.MapColor;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.block.enums.Instrument;
@@ -18,19 +19,31 @@ import net.minecraft.registry.Registry;
 import net.minecraft.sound.BlockSoundGroup;
 
 public class ModBlocks {
+    public static final AbstractBlock.Settings DETECTOR_SETTINGS = FabricBlockSettings.create()
+            .mapColor(MapColor.OAK_TAN)
+            .instrument(Instrument.BASS)
+            .strength(0.2F)
+            .sounds(BlockSoundGroup.WOOD)
+            .burnable();
+
     public static final RestartDetectorBlock RESTART_DETECTOR =
             Registry.register(Registries.BLOCK, RestartDetector.id("restart_detector"),
-                    new RestartDetectorBlock(FabricBlockSettings.create()
-                            .mapColor(MapColor.OAK_TAN)
-                            .instrument(Instrument.BASS)
-                            .strength(0.2F)
-                            .sounds(BlockSoundGroup.WOOD)
-                            .burnable()));
+                    new RestartDetectorBlock(DETECTOR_SETTINGS));
+    public static final TpsDetectorBlock TPS_DETECTOR =
+            Registry.register(Registries.BLOCK, RestartDetector.id("tps_detector"),
+                    new TpsDetectorBlock(DETECTOR_SETTINGS));
+
     public static final BlockItem RESTART_DETECTOR_ITEM =
             Registry.register(Registries.ITEM, RestartDetector.id("restart_detector"),
                     new PolymerBlockItem(RESTART_DETECTOR, new FabricItemSettings(), Items.DAYLIGHT_DETECTOR));
+    public static final BlockItem TPS_DETECTOR_ITEM =
+            Registry.register(Registries.ITEM, RestartDetector.id("tps_detector"),
+                    new PolymerBlockItem(TPS_DETECTOR, new FabricItemSettings(), Items.DAYLIGHT_DETECTOR));
 
     public static void register() {
-        ItemGroupEvents.modifyEntriesEvent(ItemGroups.REDSTONE).register(group -> group.add(RESTART_DETECTOR_ITEM));
+        ItemGroupEvents.modifyEntriesEvent(ItemGroups.REDSTONE).register(group -> {
+            group.add(RESTART_DETECTOR_ITEM);
+            group.add(TPS_DETECTOR_ITEM);
+        });
     }
 }