+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added the TPS Detector, a block that outputs the TPS and MSPT of the server as a redstone signal.
 - Slight changes to spinny command block animation, it'll now spin faster the closer it is to a restart.
   - It should also be a little smoother.
--
+- Added a new command, `/restartdetector`, with a few subcommands to use the mod even without hijacking the vanilla stop command.